+++ tests/test_contract_functionality.py
@@ -7,6 +7,7 @@ from pons import (
     Constructor,
     ContractABI,
     DeployedContract,
+    EVMVersion,
     Method,
     Mutability,
     abi,
@@ -17,7 +18,7 @@ from pons import (
 @pytest.fixture
 def compiled_contracts():
     path = Path(__file__).resolve().parent / "TestContractFunctionality.sol"
-    return compile_contract_file(path)
+    return compile_contract_file(path, evm_version=EVMVersion.CANCUN)
 
 
 async def test_empty_constructor(session, root_signer, compiled_contracts):
@@ -161,12 +162,16 @@ async def test_complicated_event(session, root_signer, compiled_contracts):
 
     basic_contract = compiled_contracts["Test"]
 
-    inner1 = [b"0123", b"1111111111"]
-    inner2 = [b"-123", b"2222222222"]
-    foo = [b"4567", [b"aa", b"bb"], b"444444", [b"0123", b"3333333333"]]
-    event_filter = basic_contract.abi.event.Complicated(b"aaaa", b"55555555", foo, [inner1, inner2])
+    bytestring33len1 = b"012345678901234567890123456789012"
+    bytestring33len2 = b"-12345678901234567890123456789012"
+    inner1 = [b"0123", bytestring33len1]
+    inner2 = [b"-123", bytestring33len2]
+    foo = [b"4567", [b"aa", b"bb"], bytestring33len1, "\u1234\u1212", inner1]
+    event_filter = basic_contract.abi.event.Complicated(
+        b"aaaa", bytestring33len2, foo, [inner1, inner2]
+    )
 
-    contract = await session.deploy(root_signer, basic_contract.constructor(123, 456))
+    contract = await session.deploy(root_signer, basic_contract.constructor(12345, 56789))
 
     log_filter1 = await session.eth_new_filter(event_filter=event_filter)  # filter by topics
     log_filter2 = await session.eth_new_filter()  # collect everything